changed city iata code file name to request_model
